+++ plone/app/contenttypes/tests/robot/variables.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 from pkg_resources import resource_filename
 
-TEST_FOLDER_ID = 'robot-test-folder'
 PATH_TO_TEST_FILES = resource_filename("plone.app.contenttypes.tests", "")